Add patch to switch default configuration to Apache 2.4 style authorization
This also disables the AuthGroupFile directive in the default
configuration (previously was necessary IIRC) and hence removes the
need to enable mod_authz_groupfile. "Require group" is commented out
by default by upstream anyway, so this doesn't hurt at all.

Since the default configuration contains a lot of RewriteRule
statements for the /hobbit/ to /xymon/ migration, enable mod_rewrite
in the postinst script automatically. (An alternative would be to add
a <IfModule rewrite> container around them.)