Add patch to workaround changed "ntpdate -p 2" behaviour
In Squeeze, "ntpdate -p 2 host" exits nearly immediately if the host
is reachable and longer if not. On Wheezy, this command always takes
at least two seconds, and hence slows down xymonnet runs immensely.
"ntpdate -p 1 host" works in Wheezy as fast as in Squeeze, though.

This patch changes the default for -p from 2 to 1 and changes its
location from being a hardcoded value in xymonnet.c to a configurable
value in xymonserver.cfg.

Closes: #776664